# backend-go/api/openapi.yaml
openapi: 3.1.0
info:
  title: EduPilot Gateway API
  version: 0.14.0
  description: |
    Hợp đồng API của gateway Go — NGUỒN SỰ THẬT (D45). Mỗi route trong mã phải có ở đây và ngược lại
    (`internal/contract` kiểm hai chiều). Route thử nằm ở `openapi.test.yaml`, chỉ có trong binary dựng bằng
    build tag `testroutes`. Phase PG chưa có endpoint nghiệp vụ; P1 (FEAT-llm-gateway US-P1-04) thêm 13 thao tác `/admin/llm/*` và `/courses/{id}/llm-budget`.
    P2 (FEAT-account-security US-P2-02) thêm 9 thao tác `/auth/{login,refresh,logout,register,verify-email,resend-verification,forgot-password,reset-password,tokens/preview}` và 4 thao tác `/me/{password,sessions,sessions/{id}}` (US-P2-04); US-P2-05 thêm 429 (`LOGIN_THROTTLED` | `RATE_LIMITED`) cho login / register / verify-email / reset-password / refresh: công khai (không Bearer); `refresh`/`logout` dùng cookie `ep_rt` (HttpOnly) + kiểm Origin.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/events:
    get:
      operationId: streamEvents
      tags: [events]
      summary: Luồng sự kiện thời gian thực (SSE) của chính người gọi
      description: |
        `text/event-stream`. Xác thực bằng header `Authorization` (không nhận token qua query). Khung sự kiện
        `id/event/data`; heartbeat `: hb` mỗi 25 s; tối đa 2 kết nối mỗi người; tự đóng sau 120 s bằng
        `event: reconnect`. Nối lại bằng header `Last-Event-ID`.
      parameters:
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Luồng SSE đang mở
          headers:
            Cache-Control:
              schema:
                type: string
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/admin/llm/routes:
    get:
      operationId: listLLMRoutes
      tags: [llm]
      summary: Tuyến của 7 tác vụ + khối embedding (ADMIN, TEACHER)
      responses:
        "200":
          description: Tuyến
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMRoutes"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      operationId: putLLMRoute
      tags: [llm]
      summary: Đặt chuỗi mô hình của MỘT tác vụ (ADMIN); hoán đổi nguyên tử và nạp lại nóng
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LLMRouteInput"
      responses:
        "200":
          description: Tuyến mới
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMRoutePut"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Chưa đăng nhập, token hết hạn / không hợp lệ / phiên bị thu hồi (`UNAUTHENTICATED`, `TOKEN_EXPIRED`, `TOKEN_INVALID`, `SESSION_REVOKED`, `INVALID_CREDENTIALS`)
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Sai vai trò, ngoài lớp hoặc Origin lạ (`FORBIDDEN`, `details.reason` = `role` | `course` | `origin`); tài khoản bị khoá (`ACCOUNT_DISABLED`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Conflict:
      description: Xung đột trạng thái hoặc phiên bản (`CONFLICT`, `VERSION_CONFLICT`, `IDEMPOTENCY_IN_PROGRESS`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationFailed:
      description: Dữ liệu không hợp lệ (`VALIDATION_FAILED`, `INVALID_CURSOR`, `IDEMPOTENCY_KEY_*`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: Quá hạn mức (`RATE_LIMITED`, `SSE_LIMIT_REACHED`); kèm `Retry-After`
      headers:
        Retry-After:
          required: true
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    LLMRoute:
      type: object
      required: [task, lane, chain, params, version]
      additionalProperties: false
      properties:
        task:
          type: string
          enum: [CHAT, CLASSIFY, UTILITY, GRADING, QUESTION_GEN, INSIGHT, EMBEDDING]
        lane:
          type: string
          enum: [INTERACTIVE, NEAR_REALTIME, BATCH]
        chain:
          type: array
          items:
            type: object
            required: [model_id, provider_id, provider_name, model]
            additionalProperties: false
            properties:
              model_id:
                type: string
                format: uuid
              provider_id:
                type: string
                format: uuid
              provider_name:
                type: string
              model:
                type: string
        params:
          type: object
          additionalProperties:
            type: number
        version:
          type: integer
          minimum: 0
    LLMRoutePut:
      type: object
      required: [task, lane, chain, params, version, reindex_required]
      additionalProperties: false
      properties:
        task:
          type: string
          enum: [CHAT, CLASSIFY, UTILITY, GRADING, QUESTION_GEN, INSIGHT, EMBEDDING]
        lane:
          type: string
          enum: [INTERACTIVE, NEAR_REALTIME, BATCH]
        chain:
          type: array
          items:
            type: object
            required: [model_id, provider_id, provider_name, model]
            additionalProperties: false
            properties:
              model_id:
                type: string
                format: uuid
              provider_id:
                type: string
                format: uuid
              provider_name:
                type: string
              model:
                type: string
        params:
          type: object
          additionalProperties:
            type: number
        version:
          type: integer
          minimum: 0
        reindex_required:
          type: boolean
    LLMRoutes:
      type: object
      required: [items, embedding]
      additionalProperties: false
      properties:
        items:
          type: array
          minItems: 7
          maxItems: 7
          items:
            $ref: "#/components/schemas/LLMRoute"
        embedding:
          type: [object, "null"]
          required: [model_id, provider_name, model, dims, reindex_required, indexed_chunks]
          additionalProperties: false
          properties:
            model_id:
              type: string
              format: uuid
            provider_name:
              type: string
            model:
              type: string
            dims:
              type: integer
            reindex_required:
              type: boolean
            indexed_chunks:
              type: [integer, "null"]
    LLMRouteInput:
      type: object
      required: [task, chain, version]
      additionalProperties: false
      properties:
        task:
          type: string
        chain:
          type: array
          items:
            type: string
            format: uuid
        params:
          type: object
          additionalProperties:
            type: number
        version:
          type: integer
          minimum: 0
...
    Error:
      type: object
      required: [code, message, trace_id]
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
            - BAD_REQUEST
            - UNAUTHENTICATED
            - TOKEN_EXPIRED
            - TOKEN_INVALID
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - VERSION_CONFLICT
            - IDEMPOTENCY_IN_PROGRESS
            - PAYLOAD_TOO_LARGE
            - UNSUPPORTED_MEDIA_TYPE
            - VALIDATION_FAILED
            - INVALID_CURSOR
            - IDEMPOTENCY_KEY_REUSED
            - IDEMPOTENCY_KEY_REQUIRED
            - RATE_LIMITED
            - SSE_LIMIT_REACHED
            - INTERNAL
            - SERVICE_UNAVAILABLE
            - NOT_READY
            - DEADLINE_EXCEEDED
            - OVERLOADED
            - LLM_NOT_CONFIGURED
            - LLM_UNAVAILABLE
            - PROVIDER_IN_USE
            - MODEL_DIMS_MISMATCH
            - ROUTE_INVALID
            - INVALID_CREDENTIALS
            - LOGIN_THROTTLED
            - ACCOUNT_DISABLED
            - SESSION_REVOKED
            - LINK_INVALID
            - EMAIL_NOT_VERIFIED
            - JOIN_CODE_INVALID
            - COURSE_FULL
            - COURSE_ARCHIVED
            - EXAM_NOT_OPEN
            - ATTEMPT_ALREADY_SUBMITTED
            - ATTEMPT_CLOSED
            - ATTEMPT_OTHER_TAB
            - DRAFT_CONFLICT
            - EXAM_LOCKED
            - QUESTION_IN_USE
            - RESULT_NOT_PUBLISHED
            - SUBMISSION_LIMIT_REACHED
            - JUDGE_UNAVAILABLE
            - APPEAL_WINDOW_CLOSED
            - APPEAL_EXISTS
        message:
          type: string
        trace_id:
          type: string
          pattern: "^[0-9a-f]{32}$"
        details: {}
        retry_after:
          type: integer
          minimum: 1
...